pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "A voice recording and transcription package."
authors = [
    { name = "Michal Michalski", email = "michal.michalski@fractile.io" }
]
readme = "README.md"
requires-python = ">=3.6"
license = { text = "MIT" }

dependencies = [
    "openai-whisper>=1.0",
    "pyaudio>=0.2.11",
    "numpy>=1.19.5"
]

[tool.setuptools.packages.find]
where = ["."]
